move utils in root folder
