+++ browserslist
@@ -1,5 +1,5 @@
 # Browsers that we support
 
 > 2%
-Last 2 versions
+Last 1 versions
 not ie > 0
