+++ gatsby-node/index.js
@@ -1,5 +1,6 @@
 exports.createPages = async gatsby => {
   await require('./podcast')(gatsby);
-  await require('./person')(gatsby);
+  await require('./persons')(gatsby);
+  // eslint-disable-next-line
   return Promise.resolve();
 };
