+++ utils/to-snake-case.js
@@ -1,0 +1,6 @@
+export function toSnakeCase(text) {
+  const lower = text.toLowerCase()
+  const snake = lower.replace(/\s/g, '_')
+
+  return snake
+}
